Refactor PayPal admin routes and improve transaction filtering
Moved PayPal admin endpoints to a dedicated router and updated backend route registration. Enhanced transaction and subscription filtering by fund name and improved error handling. Updated models to include fund_name field. Adjusted frontend and API paths to match new backend structure and endpoints. Added debugging output to subscription confirmation page.
