+++ backend/models/donation_subscription.py
@@ -16,6 +16,7 @@ class DonationSubscription(BaseModel):
     start_time: Optional[str] = None
     next_billing_time: Optional[str] = None
     name: Optional[str] = None  # Donor name
+    fund_name: Optional[str] = None  # Donation fund purpose
     note: Optional[str] = None
     currency: Optional[str] = "USD"
     event_type: Optional[str] = None
