+++ frontend/app/.env.example
@@ -2,4 +2,4 @@ CLOUDINARY_API_KEY=your_api_key_here
 CLOUDINARY_CLOUD_NAME=your_cloud_name
 CLOUDINARY_UPLOAD_PRESET=user_avatars
 
-BACKEND_URL=http://10.0.2.2:8000/api/v1
+BACKEND_URL=http://10.0.2.2:8000/
